+++ HHKCurrMechanism.cuh
@@ -35,14 +35,12 @@ extern __device__ __constant__ struct HHKCurrMechanismClass* HHKCurrMechanismCla
 //! On-GPU reference pointer to Mechanism function implementation
 extern __device__ mech_fun_t HHKCurrMechanism_mech_fxn_t;
 
-extern __device__ double HHKCurrMechanism_cuda_mech_fun(
-    void* _self,
-	void* pre_comp,
-	void* post_comp,
-	const double dt,
-	const double global_time,
-	const unsigned int curr_step
-	);
+extern __device__ double HHKCurrMechanism_cuda_mech_fun(void* _self,
+                                                        void* pre_comp,
+                                                        void* post_comp,
+                                                        const double dt,
+                                                        const double global_time,
+                                                        const uint64_t curr_step);
 
-#endif
-#endif
+#endif /* CUDA */
+#endif /* HHKCURRMECHANISM_CUH */
